Added an option to use the tickit synchrotron device, removed the _gating_permitted_in_machine_mode function so the gate will work in synchrotron modes other than User and Special #224
After inspection of the code, the method works fine. The problem is in the suspenders.py method __make_event. The default time to wait for thread safety is 0.1 seconds. In this incremement thread safety isn't achieved, and so self._ev is never set. Setting the wait time to 1 second fixes this